Updated for Alolan forms
Updated the spreadsheet to support Alolan forms. Other modifications include the addition of Gender, Fast Move and Charge Move columns to track this information. I have also added an "Index" column as the first column of the sheet, which allows you to sort based on various information (CP, Level, etc.) and return to a "default" sort (ie, matching your "Recent" sorting in game) when you're done. I also modified the code to allow the distinction between a normal "Evolve" and a "Mass Evolve" in the Remarks column to better plan out evolve sessions.

The evolver tab will now Bold & Italic pokemon which you do not have in your storage, giving you a visual indication on ones you could catch in the near future to update your candy count for them.

Expanded the storage box to the current max in game of 1500.